requirements.txt
python-nmap>=0.7.1
scapy>=2.5.0
requests>=2.32.0
